convert drawing to Rust; server work; dev plane.rs

+++ crate/src/tower.rs
@@ -2,20 +2,20 @@ use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen]
 pub struct Tower {
-    x: i32,
-    y: i32,
+    x: f64,
+    y: f64,
 }
 
 #[wasm_bindgen]
 impl Tower {
-    pub fn new(x: i32, y: i32) -> Self {
+    pub fn new(x: f64, y: f64) -> Self {
         Self { x, y }
     }
 
-    pub fn x(&self) -> i32 {
+    pub fn x(&self) -> f64 {
         self.x
     }
-    pub fn y(&self) -> i32 {
+    pub fn y(&self) -> f64 {
         self.y
     }
 }
